Move news-flash views into separate views sub-package
